requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
scipy>=1.7.0
matplotlib>=3.4.0
PyYAML>=5.4.0
scikit-image>=0.18.0
Pillow>=8.3.0
pandas>=1.3.0
